Profile: refactors component and images to profile/ folder and updates import references accordingly
