+++ src/components/about/profile/img/head-shot.png
